refactor(tokencache): extract token cache to neutral providers/tokencache package
Move TokenExchangeCache / CachedExchangeToken / TokenExchangeCacheStats /
NewTokenExchangeCache / NewTokenExchangeCacheWithMaxEntries / GenerateCacheKey
from providers/oidc/exchange.go to providers/tokencache, renamed to the
provider-neutral Cache / Token / Stats / New / NewWithMaxEntries /
GenerateCacheKey. The oidc package no longer owns or exposes the cache;
any credential provider (OIDC exchange, GitHub App installation tokens, etc.)
can import tokencache directly.

Breaking change: callers that imported these symbols from providers/oidc must
update to providers/tokencache. No deprecation aliases are provided.
