+++ CHANGELOG.md
@@ -21,11 +21,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - **Stale issuer JWKS broke every token after an issuer signing-key rotation.** `oidc.JWKSClient` cached a trusted issuer's JWKS for the full TTL (1h) and never refetched when a token presented a `kid` absent from the cached set, so a single Dex key rotation made `ValidateIDToken` (and therefore the token-exchange broker's subject-token validation) reject **every** current user token until the process was restarted ([muster#847](https://github.com/giantswarm/muster/issues/847)). The client now forces a single bounded refetch on an unknown `kid` and retries verification once. The refetch is rate-limited to at most one network fetch per `DefaultJWKSRefetchBackoff` (1m) per JWKS URI (negative caching), so a flood of tokens carrying genuinely bogus kids cannot hammer the issuer's JWKS endpoint. A legitimate rotation now self-heals on the first affected token without a restart. New exported sentinel `oidc.ErrKeyNotFound`.
 
+### Changed
+
+- `providers/oidc.TokenExchangeCache`, `CachedExchangeToken`, `TokenExchangeCacheStats`, `NewTokenExchangeCache`, `NewTokenExchangeCacheWithMaxEntries`, and `GenerateCacheKey` have moved to the new `providers/tokencache` package as `Cache`, `Token`, `Stats`, `New`, `NewWithMaxEntries`, and `GenerateCacheKey`. The old names are removed; update import paths.
+
 ### Added
 
 - `server.ExchangerRequest` now carries `ActorToken string`, `ActorTokenType string`, and `Actor *server.SubjectIdentity`. The brokered token-exchange endpoint accepts RFC 8693 `actor_token` and `actor_token_type` form parameters; when present the actor token is validated against the server's trusted issuers (same validator path as the subject token) and the verified actor identity is forwarded to the `Exchanger`. Absent actor params leave `Actor` nil and `BrokerExchangeSubjectToken` behaves identically to before.
 
 - `server.TokenExchangeUnsupportedTypeError` now exposes a `Role() string` method returning `"subject"` or `"actor"` to identify which token in the exchange request had an unsupported type. Handlers and callers that previously matched this error via `errors.As` now receive accurate role information for error messages and logs.
+
+- `providers.tokencache`: new package with a generic LRU-bounded token cache (`Cache`) for short-lived access tokens, usable by any credential provider (OIDC exchange, GitHub App installation tokens, etc.).
 
 - `server.TrustedIssuer.AcceptedTypHeaders`: per-issuer list of accepted JWT `typ` header values for Bearer validation. Empty keeps the RFC 9068 §4 default (`at+jwt`). Kubernetes ServiceAccount tokens carry no `typ` header, so the previously unconditional `at+jwt` requirement made the documented K8s SA trust use case impossible; configure `[""]` (optionally with `"JWT"`) on the cluster issuer entry to accept them. Signature, issuer, audience, and claim checks are unchanged.
